+++ binary-data-schema/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Split up error types into validation, encoding and decoding.
 - CLI tool is now separated from the crate.
 - `"minLength"` and `"maxLength"` are now related to the decoded JSON string.
+- Feature `"const"` is replaced by `"default"` as this is closer to the implementation in BDS.
 
 ### Features